feat: :art: Improved code implementation, speparation of concerns and changed function names to be more informative.
